move DefaltCheckers and BasicCompositionTests to matter.testing
