+++ src/python_testing/TC_DefaultWarnings.py
@@ -35,9 +35,9 @@
 # === END CI TEST ARGUMENTS ===
 
 from mobly import asserts
-from DefaultChecker import (FLAG_DEFAULT_CALENDAR_FORMAT, FLAG_FAULT_INJECTION, FLAG_FIXED_LABEL_DEFAULT_VALUES,
-                             FLAG_FIXED_LABEL_EMPTY, FLAG_PRODUCT_NAME, FLAG_SAMPLE_MEI, FLAG_UNIT_TESTING,
-                             FLAG_VENDOR_ID, FLAG_VENDOR_NAME, DefaultChecker)
+from matter.testing.default_checker import (FLAG_DEFAULT_CALENDAR_FORMAT, FLAG_FAULT_INJECTION, FLAG_FIXED_LABEL_DEFAULT_VALUES,
+                                            FLAG_FIXED_LABEL_EMPTY, FLAG_PRODUCT_NAME, FLAG_SAMPLE_MEI, FLAG_UNIT_TESTING,
+                                            FLAG_VENDOR_ID, FLAG_VENDOR_NAME, DefaultChecker)
 
 from matter.testing.basic_composition import BasicCompositionTests
 from matter.testing.decorators import async_test_body
