+++ src/python_testing/test_testing/TestSpecParsingDeviceType.py
@@ -14,12 +14,9 @@
 #    See the License for the specific language governing permissions and
 #    limitations under the License.
 #
-import os
 import re
-import sys
 import xml.etree.ElementTree as ElementTree
 
-sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
 from DeviceConformanceTests import DeviceConformanceTests, get_supersets
 from fake_device_builder import create_minimal_dt
 from jinja2 import Template
@@ -466,7 +463,6 @@ class TestSpecParsingDeviceType(DeviceConformanceTests):
         one_four_two, one_four_two_problems = build_xml_device_types(PrebuiltDataModelDirectory.k1_4_2)
         one_five, one_five_problems = build_xml_device_types(PrebuiltDataModelDirectory.k1_5)
         one_five_one, one_five_one_problems = build_xml_device_types(PrebuiltDataModelDirectory.k1_5_1)
-        one_six, one_six_problems = build_xml_device_types(PrebuiltDataModelDirectory.k1_6)
         self.problems.extend(one_two_problems)
         self.problems.extend(one_three_problems)
         self.problems.extend(one_four_problems)
@@ -474,7 +470,6 @@ class TestSpecParsingDeviceType(DeviceConformanceTests):
         self.problems.extend(one_four_two_problems)
         self.problems.extend(one_five_problems)
         self.problems.extend(one_five_one_problems)
-        self.problems.extend(one_six_problems)
 
         asserts.assert_equal(len(one_two_problems), 0, "Problems found when parsing 1.2 spec")
         asserts.assert_equal(len(one_three_problems), 0, "Problems found when parsing 1.3 spec")
@@ -483,7 +478,6 @@ class TestSpecParsingDeviceType(DeviceConformanceTests):
         asserts.assert_equal(len(one_four_two_problems), 0, "Problems found when parsing 1.4.2 spec")
         asserts.assert_equal(len(one_five_problems), 0, "Problems found when parsing 1.5 spec")
         asserts.assert_equal(len(one_five_one_problems), 0, "Problems found when parsing 1.5.1 spec")
-        asserts.assert_equal(len(one_six_problems), 0, "Problems found when parsing 1.6 spec")
 
         # Current ballot has a bunch of problems related to IDs being allocated for closures and TBR. These should all
         # mention ID-TBD as the id, so let's pull those out for now and make sure there are no UNKNOWN problems.
@@ -517,8 +511,6 @@ class TestSpecParsingDeviceType(DeviceConformanceTests):
                              set(), "There are some 1.4.2 device types that are unexpectedly not included in the 1.5 spec")
         asserts.assert_equal(set(one_five.keys())-set(one_five_one.keys()),
                              set(), "There are some 1.5 device types that are unexpectedly not included in the 1.5.1 spec")
-        asserts.assert_equal(set(one_five_one.keys())-set(one_six.keys()),
-                             set(), "There are some 1.5.1 device types that are unexpectedly not included in the 1.6 spec")
 
     @run_against_all_spec_revisions
     def test_application_device_type_on_root(self):
